Adding support for parallel tree building
Currently only OpenMP is supported. According to profiling computing the
split energy is the most costly operation. A simple counter measure is
to parallelize computation of split energies for all splits at a given
node.

However, I fear this kind of parallelization is not optimal in terms of
overhead. Better idea: as we walk the tree in BFS we could compute the
split energies for all nodes on the current level in parallel once, then
split all nodes and dive to the next level.