Adding parameter that controls random sample area
It appears that contour landmarks are not matched very well. This could
be improved by allowing the expansion of the sampling area for pixel
locations. Up to now this was limited to AABB of the mean shape. The
value expansionRandomPixelCoordinates now allows an offset to be
applied to the aabb min and max corner. Note that sampling is done in
normalized shape space, so values should range somewhere between 0 and 1